[Refactor] Split subscription usage provider adapters (#841)
